+++ src/router/httpd/modules/upgrade_rb500.c
@@ -210,7 +210,7 @@ err:
 
 static void
 // do_upgrade_post(char *url, FILE *stream, int len, char *boundary)
-do_upgrade_post(char *url, webs_t stream, int len, char *boundary)	// jimmy, 
+do_upgrade_post(char *url, webs_t stream, size_t len, char *boundary)	// jimmy, 
 									// https, 
 									// 8/6/2003
 {
